requirements.txt
anyio==3.6.2
asgiref==3.6.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.2
django-cors-headers==3.14.0
django-environ==0.10.0
django-filter==23.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
google-api-python-client==1.7.12
google-auth==2.17.3
google-auth-httplib2==0.1.0
h11==0.14.0
httpcore==0.17.0
httplib2==0.22.0
httpx==0.24.0
idna==3.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyJWT==2.6.0
pyparsing==3.0.9
pytz==2023.3
requests==2.28.2
rsa==4.9
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.3
uritemplate==3.0.1
urllib3==1.26.15
